# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reasoning/assessments/{assessmentId}:
    get:
      operationId: getReasoningAssessment
      tags: [reasoning]
      summary: Get one assessment with its items (no scoring keys exposed)
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAssessment"
  /reasoning/assessments/{assessmentId}/start:
    post:
      operationId: startReasoningAttempt
      tags: [reasoning]
      summary: Start (or resume) an attempt on a diagnostic assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartReasoningBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAttemptState"
  /reasoning/assessments/{assessmentId}/submit:
    post:
      operationId: submitReasoningAttempt
      tags: [reasoning]
      summary: Submit responses; records a pass and returns written feedback
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReasoningBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningResult"
...
components:
  schemas:
...
    ReasoningItem:
      type: object
      required: [id, position, type, prompt]
      properties:
        id: { type: integer }
        position: { type: integer }
        type:
          type: string
          enum: [mcq, open]
        prompt: { type: string }
        options:
          type: ["array", "null"]
          items: { type: string }
          description: For mcq items — the answer choices.
    ReasoningAssessment:
      type: object
      required: [id, instrument, phase, title, instructions, items]
      properties:
        id: { type: integer }
        instrument:
          type: string
          enum: [subject, reasoning]
        phase:
          type: string
          enum: [before, third1, third2, after]
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        instructions: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningItem"
    ReasoningAttemptState:
      type: object
      required: [id, assessmentId, status, startedAt, items]
      properties:
        id: { type: integer }
        assessmentId: { type: integer }
        status:
          type: string
          enum: [in_progress, submitted]
        format:
          type: ["string", "null"]
          description: The answer format used for this attempt (mcq | hybrid | written).
        length:
          type: ["string", "null"]
          description: The length tier used for this attempt (short | medium | long).
        startedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
        passed: { type: ["boolean", "null"] }
        feedback: { type: ["string", "null"] }
        headline:
          type: ["string", "null"]
          description: For a submitted attempt being reviewed — the score headline.
        metrics:
          type: ["array", "null"]
          description: For a submitted attempt being reviewed — the score metrics.
          items:
            $ref: "#/components/schemas/ReasoningMetric"
        review:
          type: ["array", "null"]
          description: For a submitted attempt being reviewed — per-question review with the student's answer and the correct answer.
          items:
            $ref: "#/components/schemas/ReasoningReviewItem"
        items:
          type: array
          description: >-
            The exact items to present for THIS attempt. The first take uses the
            seeded template; each retake returns freshly generated questions of
            the same kind (same instrument, skill areas, and structure).
          items:
            $ref: "#/components/schemas/ReasoningItem"
    ReasoningResponseInput:
      type: object
      required: [itemId]
      properties:
        itemId: { type: integer }
        text:
          type: ["string", "null"]
          description: open — the student's short typed answer (1-2 sentences).
        selectedIndex:
          type: ["integer", "null"]
          description: mcq — chosen option index.
...
    StartReasoningBody:
      type: object
      properties:
        retake:
          type: boolean
          description: >-
            When true, begin a fresh attempt even if a previous attempt was
            already submitted. An in-progress attempt is still resumed.
        format:
          type: string
          enum: [mcq, hybrid, written]
          description: >-
            The answer format the student picked for this attempt. "mcq" =
            multiple choice only (no typing); "hybrid" = mostly multiple choice
            plus 1-2 one-sentence written answers; "written" = short open
            answers only. Applied only when creating a new attempt; ignored when
            resuming or reviewing. Defaults to "mcq".
        length:
          type: string
          enum: [short, medium, long]
          description: >-
            How many questions the attempt should contain. "short" = a few
            questions, "medium" = a moderate number, "long" = many questions.
            Independent of format. Applied only when creating a new attempt;
            ignored when resuming or reviewing. Defaults to "medium".
    SubmitReasoningBody:
      type: object
      required: [responses]
      properties:
        responses:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningResponseInput"
    ReasoningMetric:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
        detail: { type: ["string", "null"] }
    ReasoningReviewItem:
      type: object
      required: [itemId, type, prompt]
      properties:
        itemId: { type: integer }
        type:
          type: string
          enum: [mcq, open]
        prompt: { type: string }
        options:
          type: ["array", "null"]
          items: { type: string }
          description: mcq — the answer choices shown.
        selectedIndex:
          type: ["integer", "null"]
          description: mcq — the option index the student chose.
        correctIndex:
          type: ["integer", "null"]
          description: mcq — the correct option index.
        isCorrect:
          type: ["boolean", "null"]
          description: mcq — whether the student's choice was correct.
        text:
          type: ["string", "null"]
          description: open — the student's typed answer.
        expectedPoints:
          type: ["array", "null"]
          items: { type: string }
          description: open — the key idea(s) a good brief answer should capture.
        rationale:
          type: ["string", "null"]
          description: open — the grader's brief explanation of correctness.
    ReasoningResult:
      type: object
      required: [attemptId, passed, feedback, headline, metrics, review]
      properties:
        attemptId: { type: integer }
        passed: { type: boolean }
        feedback: { type: string }
        headline: { type: string }
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningMetric"
        review:
          type: array
          description: Per-question review — each item with the student's answer and the correct answer.
          items:
            $ref: "#/components/schemas/ReasoningReviewItem"
...