# lib/api-zod/src/generated/types/reasoningResponseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningResponseInput {
  itemId: number;
  /**
     * open — the student's short typed answer (1-2 sentences).
     * @nullable
     */
  text?: string | null;
  /**
     * mcq — chosen option index.
     * @nullable
     */
  selectedIndex?: number | null;
}