# lib/api-zod/src/generated/types/reasoningReviewItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasoningReviewItemType } from './reasoningReviewItemType';

export interface ReasoningReviewItem {
  itemId: number;
  type: ReasoningReviewItemType;
  prompt: string;
  /**
     * mcq — the answer choices shown.
     * @nullable
     */
  options?: string[] | null;
  /**
     * mcq — the option index the student chose.
     * @nullable
     */
  selectedIndex?: number | null;
  /**
     * mcq — the correct option index.
     * @nullable
     */
  correctIndex?: number | null;
  /**
     * mcq — whether the student's choice was correct.
     * @nullable
     */
  isCorrect?: boolean | null;
  /**
     * open — the student's typed answer.
     * @nullable
     */
  text?: string | null;
  /**
     * open — the key idea(s) a good brief answer should capture.
     * @nullable
     */
  expectedPoints?: string[] | null;
  /**
     * open — the grader's brief explanation of correctness.
     * @nullable
     */
  rationale?: string | null;
}